move modals refactoring
