+++ view/template/modal/modal_add_device.php
