+++ view/template/modal/modal_admin_device.php
