+++ view/template/modal/modal_delete_device.php
