+++ view/template/modal/modal_history_device.php
